completing pomodoro maintenance

+++ js/commands/sticker.js
@@ -1,4 +1,4 @@
-const { getStickerDB, getTodoDB, updateTodoDB } = require('../db.js')
+const { getStickerDB, getTodoDB, updateTodoDB, getVipDB } = require('../db.js')
 const { MessageEmbed } = require('discord.js')
 
 let prefix = process.en
@@ -7,15 +7,20 @@ module.exports = {
 	name: 'sticker',
 	description: 'Memilih template sticker to do list yang akan digunakan untuk menunjukan status to do list (default, onGoing, done, fail)',
 	aliases: [ 'tem', 'tpl', 'sticker' ],
-	usages: [ ',p template', ',p template <id>' ],
-	examples: [ ',p template', ',p tpl 2En75A' ],
+	usages: [ `${prefix} template`, `${prefix} template <id>`],
+	examples: [ `${prefix} template`, `${prefix} tpl 2En75A` ],
   async execute(msg, args) {
 		const templates = await getStickerDB(msg.guild.id)
 		const templatesMap = new Map(templates)
-
+		const vipRoles = await getVipDB(msg.guild.id)
 		if (templates.length < 1) return msg.channel.send('Template Server Kosong')
 
+		const userRoles = msg.member.roles.cache.map(role => role.id)
 		const template = templatesMap.get(args[0])
+
+		const isUserVip = vipRoles.some(role => {
+			return userRoles.some(userRole => userRole === role)
+		})
 
 		console.log(template)
 
@@ -23,12 +28,15 @@ module.exports = {
 			const reacts = [ '❌',  '✅' ]
 			const filter = (reaction, user) => reacts.some(react => react === reaction.emoji.name) && user.id === msg.author.id	
 			const todoOptions = [ 'default', 'on Going', 'Completed (done)', 'Uncompleted (fail)' ]
+			
+			let vipString = '🔹'
+			if (template.vip) vipString = '💎'
 
 			const templateEmbed = new MessageEmbed()
 				.setColor('#73cfff')
 				.setTitle(`${template.name} Templates`)
 				.setThumbnail(msg.guild.iconURL())
-				.setDescription(`\n> **ID**  : \`${args[0]}\`\n> **Stiker** : \n ${template.sticker.map((sticker, index) => `>   ${sticker} : ${todoOptions[index]}`).join('\n')}`)
+				.setDescription(`\n> \`ID      \`: \`${args[0]}\`\n> \`Nama    \`: **${template.name}** (${vipString})\n> \`Stiker  \`: \n ${template.sticker.map((sticker, index) => `>   ${sticker} : ${todoOptions[index]}`).join('\n')}`)
 				.setFooter(`❌ : batal | ✅ : Gunakan Template`)
 
 			const embed = await msg.channel.send(templateEmbed)
@@ -43,9 +51,10 @@ module.exports = {
 						return
 
 					case '✅':
+						if (template.vip && !isUserVip) return msg.channel.send('Sticker ini hanya untuk member VIP')
 						getTodoDB(msg.author.id).then(todo => {
 							todo.sticker = args[0]
-							updateTodoDB(msg.author.id, Array.from(userTemplatesMap))
+							updateTodoDB(msg.author.id, todo)
 							embed.delete()
 							msg.channel.send('Template Berhasil diganti')
 						})
@@ -63,7 +72,7 @@ module.exports = {
 			.setFooter(`-/${templates.length} | ✅ : Gunakan Template | ❌ : batal`)
 
 		let container = await msg.channel.send(initialEmbed)
-		renderEmbed(msg, templates, 0, container)
+		renderEmbed(msg, templates, 0, container, isUserVip)
 		const templatesReact = [ '⬅️', '➡️', '✅', '❌'  ]
 		for (const react of templatesReact) {
 			container.react(react)
@@ -72,7 +81,7 @@ module.exports = {
 }
 
 
-async function renderEmbed(msg, templates, index, embed) {
+async function renderEmbed(msg, templates, index, embed, isUserVip) {
 	const template = templates[index]
 	const todoOptions = [ 'default', 'on Going', 'Completed (done)', 'Uncompleted (fail)' ]
 
@@ -98,17 +107,18 @@ async function awaitTemplateReaction(msg, templates, index, embed) {
 		case '⬅️':
 			newIndex--
 			if (newIndex < 0) newIndex = templates.length - 1
-			renderEmbed(msg, templates, newIndex, embed)
+			renderEmbed(msg, templates, newIndex, embed, isUserVip)
 			return
 
 		case '➡️':
 			newIndex++
 			console.log(newIndex)
 			if (newIndex >= templates.length) newIndex = 0
-			renderEmbed(msg, templates, newIndex, embed)
+			renderEmbed(msg, templates, newIndex, embed, isUserVip)
 			return
 		
 		case '✅':
+			if (templates[index].vip && !isUserVip) return msg.channel.send('Sticker ini hanya untuk member VIP')
 			let todoData = await getTodoDB(msg.author.id)
 			todoData.sticker = templates[index][0]
 			updateTodoDB(msg.author.id, todoData)
